refactor: 重构Spark模型
