+++ .gitignore
@@ -15,3 +15,4 @@ x64/
 reports/
 results/
 __pycache__/
+models/
