+++ iAgents/Spark/layers/ABCActivator.py
