--- iAgents/Spark/modules/__init__.py
@@ -1,10 +1,0 @@
-from .layers.dense import Dense
-from .layers.relu import ReLU
-
-from .loss.mae import MAE
-from .loss.mse import MSE
-from .loss.sce import SoftmaxCrossEntropy
-
-from .optimizers.adam import Adam
-
-from .net import Net
