+++ iAgents/Spark/optimizer/adagrad.py
@@ -5,7 +5,7 @@ Adagrad optimizer.
 __author__ = 'borgwang'
 __url__ = https://github.com/borgwang/tinynn
 """
-from ABCOptimizer import Optimizer
+from .ABCOptimizer import Optimizer
 
 
 class Adagrad(Optimizer):
