+++ iAgents/Spark/optimizer/monentum.py
@@ -5,7 +5,7 @@ Momentum optimizer.
 __author__ = 'borgwang'
 __url__ = https://github.com/borgwang/tinynn
 """
-from ABCOptimizer import Optimizer
+from .ABCOptimizer import Optimizer
 
 
 class Momentum(Optimizer):
